refactor: make CommitNavigationGraph route-agnostic
Part of #477
